Version 0.8.1 -Microwave fix for vessels with more than one transmitter -Improved resource manager handling of high time acceleration -Fixed the solar panel power curves and made the solar panel waste heat module and inverse square law fix apply to all solar panels, rather than just stock ones -Fixed Tritium breeding with fusion reactors -Rewrote the antimatter reactor consumption code to a custom method, which should prevent the reactor power output from ever unexpectedly dropping down to zero while you still have fuel and need power. -Changed Refinery to remove top attach node, this prevents problems with it failing to attach at certain times in the VAB -Applied radius fixes to the generator meaning that a generator works best when paired with a reactor of the right size -Made propellant definitions for the nuclear and electric engines compatible with ModuleManager -Updated RealFuels compatibility file -Tweaked fusion resource consumption rates to make them correct -Removed all the const definitions from the science lab and refinery and moved them to a seperate class, for some reason KSP sometimes takes offense to these on part loading.

+++ FNPlugin/FNResourceSuppliable.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace FNPlugin {
     interface FNResourceSuppliable {
-        void receiveFNResource(float power_supplied,String resourcename);
+        void receiveFNResource(double power_supplied,String resourcename);
         float consumeFNResource(double power_to_consume, String resourcename);
         float consumeFNResource(float power_to_consume, String resourcename);
     }
